дописал Entyty класс

+++ history/src/main/java/com/bank/history/Entities/History.java
@@ -11,11 +11,11 @@ import lombok.RequiredArgsConstructor;
 public class History {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-    private int transfer_audit_id;
-    private int profile_audit_id;
-    private int account_audit_id;
-    private int anti_fraud_audit_id;
-    private int public_bank_info_audit_id;
-    private int authorization_audit_id;
+    private long id;
+    private  long  transfer_audit_id;
+    private  long  profile_audit_id;
+    private  long  account_audit_id;
+    private  long  anti_fraud_audit_id;
+    private  long  public_bank_info_audit_id;
+    private  long  authorization_audit_id;
 }
